Upgrading HA to latest
Allegedly fixes https://github.com/home-assistant/core/issues/132541, which
sounds suspiciously similar to the behavior I'm seeing with the Christmas
tree/bannister outlets that I have set up in the SmartLife app and in HA via
Tuya. But I was previously running 2024.11, and that issue was reported with
2024.12.2 I think, so that doesn't make sense. Anyway, I'm getting new outlets
on Monday and will see if those behave better.